Reimplement the ticker and console in the new HUD framework.  Remove the old HUD framework.  Remove the old font subsystem.  Remove the Overlay component from the Ogre build.  Remove a lot of unused lua bindings.  Some things may be broken by this huge patch.  Includes Windows and Linux builds.

+++ trunk/grit_core/src/gfx/GfxFont.cpp
@@ -52,7 +52,7 @@ GfxFont *gfx_font_get (const std::string &name)
     return db[name];
 }
 
-GfxFont *gfx_font_make (const std::string &name, GfxTextureDiskResource *dr, float height)
+GfxFont *gfx_font_make (const std::string &name, GfxTextureDiskResource *dr, unsigned long height)
 {
     GfxFont *f = gfx_font_get(name);
     if (f == NULL) {
